dashboard/requirements.txt
streamlit==1.28.0
plotly==5.15.0
pandas==2.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0